[Estimator] Move estimators for BLDC rotor into their own class
Topic: estimator-move
Relative:
Reviewers:
