+++ control_loop/bldc/math_foc.hpp
@@ -3,16 +3,7 @@
 
 #include <stdint.h>
 
-#ifndef M_PI
-#define M_PI 3.14159265358979323846
-#endif
-
 namespace math {
-
-constexpr float sqrt_3 = 1.7320508075688772;
-constexpr float sqrt_3_over_3 = 0.5773502691896258;
-
-constexpr float M_PI_FLOAT = static_cast<float>(M_PI);
 
 typedef struct {
     float alpha;
